[cryptography] Update function
